Forward port SVN r1228245 to PyKDE4 git master:
Unbreak Python Phonon bindings. Now it is possible to use a VideoWidget again with a MediaSource object, which would throw a TypeError earlier (this was dueto a missing inheritance of a class). My testing confirms that it now  works.
Original patch by Carlos Corbacho (with minimal changes on my part).
Thanks!

CCBUG: 188315