Add support for konsolepart to PyKDE4 binding
In order to access the KTerminalInterface and KTerminalInterfaceV2
exposed by konsolepart, it is necessary to make use of qobject_cast.
This is not something that seems to be possible from PyQt4 and so some
C++ help is needed. This binding provides that C++ helper code, and
wrappers to match.

A sample program is also provided.
REVIEW: 101269
DIGEST: PyKDE4 get support for konsolepart.