Enable parallel installation of pykdeuic4.
Currently pykdeuic4 is installed to BINDIR indepdendently from the
Python version used for building, potentially overwriting an existing
install with a different version (think Python 2 and Python 3) and
preventing multiple installs.

This patch introduces the following changes:

- Generate a symlink from pykdeuic4.py (hosted in the site-packages dir)
  to BINDIR/pykdeuic4-VERSION_MAJOR.VERSION_MINOR, where VERSION_MAJOR
  and VERSION_MINOR are the Python major and minor version,
  respectively
- If the PYKDEUIC4_ALTINSTALL parameter is FALSE (the default) this
  symlink is further symlinked to BINDIR/pykdeuic4
- If PYKDEUIC4_ALTINSTALL is TRUE, the second symlink is not generated

Fundamentally, this is analogous to Python's own "make altinstall"

This change also allows distributions to enable parallel installs and
potentially handle the final symlink on their own (e.g. with Debian's
update-alternatives).

As PYKDEUIC4_ALTINSTALL is by default FALSE, the current behavior (prior
to this change) is preserved.