Revert "Fix build after 9b2593a"
Apparently the CMake policy is set back somewhere in CMakeLists, and some
people (not me, unfortunately) are experiencing missing symbols in
PyKDE4.

The commit (in agreement with krop) has been reverted until the real
cause is to be found.

Close the bug if you're no longer affected by this.

CCBUG: 312618

This reverts commit f8ab7037ee78746de2c30079b3e0832537fb62f7.